requirements.txt
python-dotenv>=1.0.0
httpx>=0.27.0
py-clob-client>=0.18.0
web3>=6.15.0
eth-account>=0.11.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
jinja2>=3.1.0
aiosqlite>=0.20.0
python-multipart>=0.0.12
